【********】慢卡慢链路专家建议Part2-增加前后端接口、增加慢卡/慢分组阈值（含UT）
# message auto-generated for no-merge-commit merge:
merge slowRankAdvice0604 into master

【********】慢卡慢链路专家建议Part2-增加前后端接口、增加慢卡/慢分组阈值（含UT）

Created-by: insightDeveloper

Match-id-ce93352dd20df42fad095307f82948aa57c3a917

+++ server/src/modules/summary/core/TopNAdviceMaintainer.cpp
@@ -33,7 +33,7 @@ std::vector<AdviceInfoForSlowRank> TopNAdviceMaintainer::GetTopNSlowest(uint32_t
     std::reverse(all.begin(), all.end());
 
     // 截取前 topN
-    if (topN < all.size()) {
+    if (static_cast<size_t>(topN) < all.size()) {
         all.resize(topN);
     }
     return all;
